Bugfix: supports all Github events
Github workflows that uses this action and being triggered by
`workflow_dispatch` or `schedule` events were failing. There was no
fallback on events that were not `push` or `pull-request`. The changes
in this commit make it work for all by falling back on full repo scan
whenever the event is not a `pull-request` one.

Refs:
- https://onefootball.slack.com/archives/CFUCQUA04/p1625738112043100